# packages/api-types/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisOut": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnOut"
            },
            "title": "Columns",
            "type": "array"
          },
          "data_rows": {
            "title": "Data Rows",
            "type": "integer"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldStatusOut"
            },
            "title": "Fields",
            "type": "array"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "header_row": {
            "title": "Header Row",
            "type": "integer"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "ocr": {
            "title": "Ocr",
            "type": "boolean"
          },
          "sheet": {
            "title": "Sheet",
            "type": "integer"
          },
          "sheets": {
            "items": {
              "$ref": "#/components/schemas/SheetOut"
            },
            "title": "Sheets",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "filename",
          "source",
          "ocr",
          "sheets",
          "sheet",
          "header_row",
          "columns",
          "fields",
          "data_rows",
          "notes"
        ],
        "title": "AnalysisOut",
        "type": "object"
      },
...
      "Body_analyzeImport": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "header_row": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Header Row"
          },
          "sheet": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheet"
          },
          "sheet_url": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheet Url"
          }
        },
        "title": "Body_analyzeImport",
        "type": "object"
      },
      "Body_previewImport": {
        "properties": {
          "date_order": {
            "default": "dmy",
            "enum": [
              "dmy",
              "mdy"
            ],
            "title": "Date Order",
            "type": "string"
          },
          "file": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "header_row": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Header Row"
          },
          "mapping": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mapping"
          },
          "sheet": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheet"
          },
          "sheet_url": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheet Url"
          }
        },
        "title": "Body_previewImport",
        "type": "object"
      },
...
      "ColumnOut": {
        "properties": {
          "alternatives": {
            "items": {
              "type": "string"
            },
            "title": "Alternatives",
            "type": "array"
          },
          "confidence": {
            "enum": [
              "high",
              "uncertain",
              "none"
            ],
            "title": "Confidence",
            "type": "string"
          },
          "date_info": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DateInfoOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "header": {
            "title": "Header",
            "type": "string"
          },
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "samples": {
            "items": {
              "type": "string"
            },
            "title": "Samples",
            "type": "array"
          },
          "suggested": {
            "anyOf": [
              {
                "enum": [
                  "customer_name",
                  "phone",
                  "amount_due",
                  "reference",
                  "due_date"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested"
          }
        },
        "required": [
          "index",
          "header",
          "samples",
          "suggested",
          "confidence",
          "alternatives",
          "date_info"
        ],
        "title": "ColumnOut",
        "type": "object"
      },
...
      "ConfirmIn": {
        "additionalProperties": false,
        "properties": {
          "filename": {
            "default": "",
            "maxLength": 255,
            "title": "Filename",
            "type": "string"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/RowIn"
            },
            "maxItems": 2000,
            "minItems": 1,
            "title": "Rows",
            "type": "array"
          },
          "source": {
            "anyOf": [
              {
                "enum": [
                  "excel",
                  "csv",
                  "pdf",
                  "google_sheets"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "required": [
          "rows"
        ],
        "title": "ConfirmIn",
        "type": "object"
      },
      "ConfirmOut": {
        "properties": {
          "imported": {
            "title": "Imported",
            "type": "integer"
          }
        },
        "required": [
          "imported"
        ],
        "title": "ConfirmOut",
        "type": "object"
      },
...
      "FieldStatusOut": {
        "properties": {
          "column": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Column"
          },
          "competing": {
            "items": {
              "type": "integer"
            },
            "title": "Competing",
            "type": "array"
          },
          "field": {
            "enum": [
              "customer_name",
              "phone",
              "amount_due",
              "reference",
              "due_date"
            ],
            "title": "Field",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "required": {
            "title": "Required",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "detected",
              "uncertain",
              "missing"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "field",
          "label",
          "required",
          "status",
          "column",
          "competing"
        ],
        "title": "FieldStatusOut",
        "type": "object"
      },
      "GoogleConfigOut": {
        "properties": {
          "private_access": {
            "title": "Private Access",
            "type": "boolean"
          },
          "service_account_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Account Email"
          }
        },
        "required": [
          "private_access",
          "service_account_email"
        ],
        "title": "GoogleConfigOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PreviewOut": {
        "properties": {
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "invalid": {
            "title": "Invalid",
            "type": "integer"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "ocr": {
            "title": "Ocr",
            "type": "boolean"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/RowOut"
            },
            "title": "Rows",
            "type": "array"
          },
          "skipped": {
            "items": {
              "$ref": "#/components/schemas/SkippedOut"
            },
            "title": "Skipped",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "valid": {
            "title": "Valid",
            "type": "integer"
          }
        },
        "required": [
          "filename",
          "source",
          "ocr",
          "rows",
          "skipped",
          "notes",
          "total",
          "valid",
          "invalid"
        ],
        "title": "PreviewOut",
        "type": "object"
      },
...
      "RowOut": {
        "properties": {
          "amount_due": {
            "title": "Amount Due",
            "type": "string"
          },
          "customer_name": {
            "title": "Customer Name",
            "type": "string"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "row_number": {
            "title": "Row Number",
            "type": "integer"
          },
          "warnings": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "row_number",
          "customer_name",
          "phone",
          "amount_due",
          "reference",
          "due_date",
          "errors"
        ],
        "title": "RowOut",
        "type": "object"
      },
      "RowsIn": {
        "additionalProperties": false,
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/RowIn"
            },
            "maxItems": 2000,
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "rows"
        ],
        "title": "RowsIn",
        "type": "object"
      },
...
      "SheetOut": {
        "properties": {
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rows": {
            "title": "Rows",
            "type": "integer"
          }
        },
        "required": [
          "index",
          "name",
          "rows"
        ],
        "title": "SheetOut",
        "type": "object"
      },
      "SkippedOut": {
        "properties": {
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "row_number": {
            "title": "Row Number",
            "type": "integer"
          }
        },
        "required": [
          "row_number",
          "reason"
        ],
        "title": "SkippedOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Collections API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/imports/analyze": {
      "post": {
        "operationId": "analyzeImport",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyzeImport"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read a file or Google Sheet and suggest what each column means",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports/confirm": {
      "post": {
        "operationId": "confirmImport",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "description": "Some rows still have errors"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save the reviewed rows as customers (all or nothing)",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports/google-sheets": {
      "get": {
        "operationId": "getGoogleSheetsConfig",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleConfigOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Whether private Google Sheets can be imported, and who to share them with",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports/preview": {
      "post": {
        "operationId": "previewImport",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_previewImport"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Apply the confirmed column mapping and show the normalised rows",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports/validate": {
      "post": {
        "operationId": "validateImportRows",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RowsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RowOut"
                  },
                  "title": "Response Validateimportrows",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Re-check rows after the employer edits them",
        "tags": [
          "imports"
        ]
      }
    },
...
  }
}